backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
easyocr
pandas
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain_chroma
langchain_ollama
lxml
wikipedia
pypdf
pdf2image
langchain-mistralai
spacy==3.8.7
numpy<2.0
protobuf<6.0
openai-whisper
torch
torchaudio